Add inlining tests for code considered too difficult to inline
The tests include the parts where we unconditionally consider the
function too difficult to inline, i.e. where the node count is not
taken into account. For this reason, these cases are not considered
part of the heuristics but of the inlining capabilities.

Regarding loops, I'm not sure if we disallow them by default due to
heuristics, or only allow forced inlining due to incomplete
capabilities.

Change-Id: Iee37f972c694251e0e1390fbb82bc30a4fc0bd6a
Reviewed-on: https://dart-review.googlesource.com/24109
Reviewed-by: Stephen Adams <sra@google.com>