Stop running global transformations as build transformations
Kernel transformations are categorized as 'modular' or 'global'.

Modular transformations are run at a granularity finer than a whole program
(e.g., a library, a class, or even a method).  They cannot make closed-world
assumptions.  One can think of them as running at 'compile time'.

Global transformations are run on a whole (closed) program.  They should be
optimizations only, not required for the implementation correctness of the
back end, because they are not guaranteed to be performed.  A simple linker
might just concatenate .dill files.  One can think of them as running at
'link time'.

The front end was applying global transformations to (open) bundles of
libraries, which violates the closed-world assumptions that these
transformations are allowed to make.  We no longer do that.

Bug:
Change-Id: If50faa78c3cfbcd8a3bc2f45b922c90d11761117
Reviewed-on: https://dart-review.googlesource.com/24740
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Kevin Millikin <kmillikin@google.com>