Retire front_end_inference_test.dart
While front end type inference logic was in development, this test was
useful for ensuring that the front end type inference behavior matched
the behavior of the analyzer.  Now that front end type inference
development is complete, and we are starting to make material
improvements to it, the test will only slow us down.

Change-Id: I88e6e6644bdc9bc4c953b9d484a7d9e1b8863da8
Reviewed-on: https://dart-review.googlesource.com/24643
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>