Re-enabling linter and normalizer tests.
All status files from the tests/ and runtime/tests directories has been copied
to a local directory, to make sure, that the unit test does not break when junk
is added to status files.

The linter has been updated with a check for duplicate entries in status files,
which is why the normalizer unit tests failed in the first place.

Bug:
Change-Id: Ifebd2bfd9011373bfe9a613de5c365ff99535ca4
Reviewed-on: https://dart-review.googlesource.com/25380
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Morten Krogh-jespersen <mkroghj@google.com>