[CFE/dart2js] Create dart2js-specific CFE entry point
This CL creates the initial dart2js-specific CFE entry point.
It is currently mostly a matter of copying the dart2js code for running the
front-end into the front-end, although it also applies a workaround for
making batch-mode faster.

Running the same dart2js with kernel tests as on the bots locally I get
this:

* Before: [09:11 | 100% | +10507 | -    0]
* Now [04:46 | 100% | +10507 | -    0]

Change-Id: I68f147fd7aadb0281807b4688b9f6fdc5aa24c28
Reviewed-on: https://dart-review.googlesource.com/25300
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>