Retire front_end_runtime_check_test.dart
While front end type runtime check logic was in development, this test
was useful for ensuring that the front end runtime check behavior
matched the behavior of the analyzer.  Now that front end runtime
check development is complete, and we are starting to make material
improvements to it, the test will only slow us down.

Change-Id: I1d6c6e08c809927c770ed33586f2971d6a25d936
Reviewed-on: https://dart-review.googlesource.com/26920
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>