Update dart_style to 1.0.9 and bring in its new dependencies.
This removes the dependency on scheduled_test which is important for
rolling other packages. In return, it brings in two new packages,
test_descriptor and test_process.

There are no changes in the formatter's actual output. All of the
differences are in tests or internal dependencies, so I don't think we
need to worry about coordinating this with the PRESUBMIT version of
dartfmt as in https://github.com/dart-lang/sdk/issues/30164.

Change-Id: I1d6426303055a4400a0ca3d16965a3e05ab39cf7
Reviewed-on: https://dart-review.googlesource.com/32445
Reviewed-by: Kevin Moore <kevmoo@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>