Bring dart_style 1.0.9 into the repo and update its dependencies.
This removes scheduled_test which is important for rolling other
packages. In return, it brings in two new packages, test_descriptor and
test_process.

There are no changes in the formatter's actual output. All of the
differences are in tests or internal dependencies, so I don't think we
need to worry about coordinating this with the PRESUBMIT version of
dartfmt as in https://github.com/dart-lang/sdk/issues/30164.

Change-Id: I1c81e157c32326801985615abc353f2239fe2470
Reviewed-on: https://dart-review.googlesource.com/32664
Reviewed-by: Zach Anderson <zra@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>