Don't run dart2 tests without common front end
The bare VM without the common front end (kernel) does not support Dart2
semantics, which are assumed in the corelib_2 and language_2 directories. For
now we still run the bare VM in checked mode on these tests, since that is
mostly compatible, and gives us better coverage.

R=vegorov@google.com

Change-Id: I1a1ca939d8a33503d19b683b0e5539e0fa6e9dc1
Reviewed-on: https://dart-review.googlesource.com/35160
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>
Commit-Queue: Erik Corry <erikcorry@google.com>