[VM] Use quadratic probing in hash_table.h
On my workload this reduces the probing by 19% for the CanonicalType
table, and reduces the number of insertions that take 25 probes or
more by 74%. Max load is reduced from 0.75 to 0.71.  The size is now
a power of 2 so we can use masking instead of integer division, and
we use quadratic probing in order to reduce clustering.

R=kustermann@google.com

Change-Id: I806197680a839dac99b7c7f5f09d2bce4d015235
Reviewed-on: https://dart-review.googlesource.com/34760
Commit-Queue: Erik Corry <erikcorry@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>